Bug 1256427 - Random state file and Key file should be created in $OPENSHFIT_DATA_DIR
https://bugzilla.redhat.com/show_bug.cgi?id=1256427
When creating a CSR on a gear, openssl (by default) attempts to create a 'random state' file at `$HOME/.rnd`. Users do not have write permissions on the root of their home directory, resulting in an error reported by openssl. Ensure the 'random state' file, as well as the key file, are created in the user-writable directory $OPENSHIFT_DATA_DIR.