Updated getting-started-overview.html b/c of Ruby version issue
Multiple customers on windows are reporting issues with installing the latest version of Ruby. As it turns out, the latest v2.x versions do not work. Only the latest v1.9 versions do. It is not clear on the version outside of the "Tip" which should be placed first. Hopefully, this correction removes the confusion.