Point to RubyInstaller downloads instead of archives
One of the recommended versions (2.0.0) is available on the download page, but the archive page does not have the most recent security patches available. I changed the link to point to the download page instead of the archive page. Archives can still be reached from the download page with a single click if the user prefers version 1.9.3.